Extract remaining Odoo driver CLI groups
